fix(nvim): fix recursion limit when updating title
When opening nvim with argc (not sure why), opening snacks picker for
recent and smart makes MiniIcons to hit the recursion limit when trying
to get the icon (E1169).

My suspicion is that it tries to update the title for EVERY single items
in the picker. So currently, whenever the buftype is equal to prompt /
nofile, set a default icon for it.