feat(eza): new icons
Added new icons for:
- .bash_history
- .bash_logout
- .gitconfig
- .histfile
- .node_repl_history
- .profile
- .tabiew_history
- .wget-hsts
- .wslconfig
- .zcompfump
- .zprofile
- .zshenv
- Desktop/
- Documents/
- *.appx
- *.bak
- *.cerr
- *.css
- *.dll
- *.ini
- *.kml
- *.log
- *.nss
- *.nu
- *.nu.tmpl
- *.ovpn
- *.rpd
- *.theme
- *.toml
- *.yaml
- *.yml
- *.yml.tmpl

Changed icons for:
- .bash_profile
- dot_gitconfig.tmpl
- *.yaml.tmpl

Fix override for:
- .zsh/ -> directory icon is overriden by the zsh extension, as file
still need to have a filename before the .zsh suffix, it's fine to
assign the directory glyph for this one
- copilot.lua -> same case as the above, but this time it's the lua
extension. Use the directory glyph for this one as it's one of the nvim
plugins I used. As long as I don't name a file as copilot.lua it's file
I guess..