feat(nvim): try restoring cursor whenever quitting visual mode
Problem:
A common issue with the mark approach for the select all is that
whenever we go out of visual mode (e.g. going to Command mode), the mark
stays and when we hit <ESC> we return to those marks.

So to try mitigate it, we try restoring the cursor everytime we're going
out of visual mode.